refactor(clean): estrarre read_csv_normalized.py
Separa le funzioni di lettura CSV normalizzata dal runtime duckdb:
- _normalized_csv_reader_kwargs, _load_normalized_csv_frame,
  _execute_normalized_csv_read → toolkit/clean/read_csv_normalized.py
- duckdb_read.py ora importa _execute_normalized_csv_read da read_csv_normalized
- rimossi csv e normalize_encoding (non più usati in duckdb_read.py)
- test_clean_csv_columns.py aggiornato con nuovo import

Nessun contract esterno cambiato; backward compat preservata.
